Add a version selector dropdown to the main menu
This adds a version selection dropdown to the menu. It is used to redirect on
change to the version URL. The ``versions`` theme option is added and is used to
add a mapping of version slugs to URLs.

Currently, this is going to be used in pydoc.io. The implementation will change
the javascript slightly, to accept json data instead of a blob of HTML.

This will currently be disabled on Read the Docs, as we are rethinking what the
version selection menu looks like.