Add automatic deployment to pypi when tag is pushed
`setuptools_scm` works that it provides a version even when you build it
locally and it generates the version into a file at the time of
packaging, which means there is no run-time-dependenty on the
`setuptools_scm` package later on (and also no dependency on git, or
internet access). This makes a simplified release flow, because it
enables you to just add a tag and CI pushes the end package to the PyPi
for end users to use.

Option is also to provide development packages, etc. (just the matter of
how far we want to take it). There is no required tag name, although we
should follow some of the standard formats of Major.minor.patch or
whatever it is preferred.