Fix canonical URLs with dirhtml builder
As suggested by @ericholscher in #285, using the logic from the
readthedocs-sphinx-ext package.

Currently if you set the `canonical_url` theme option (for example, to
`https://example.com/`), a canonical link like the following will be
generated for source files like `foo/bar.rst`:

```
<link rel="canonical" href="https://example.com/foo/bar.html"/>
```

This is good for the `html` builder, but doesn't match the output of the
`dirhtml` builder. This commit changes canonical link generation like
the above example to the following when the `dirhtml` builder is used:

```
<link rel="canonical" href="https://example.com/foo/bar/"/>
```

Canonical link generation remains the same as before when the `dirhtml`
builder is not used.