Release 0.5.0rc1
This is a release candidate for 0.5.0, which adds support for both HTML4
and HTML5 writers, which should support the default writer regardless of
your version of Sphinx.

Also included in this release are localization efforts and 10 language
translations.

You will need to specify the release manually to test, if there are no
major issues, we will have a 0.5.0 release shortly.

Closes #886
Closes #712