Fix a number of issues around margins on last children
The HTML4 writer depended on the `.last` selector, a class added by
Sphinx directly on output. Both writers are not using that method
anymore and instead relying on well supported CSS selectors.

This actually also addresses a number of inconsistencies with margin
spacing in lists. Version 0.4.3 and prior have a number of odd issues
around nested lists that are resolved with this change.