Drop support for sphinx 0.x and 1.x
We now depend on the html5 builder and we want to deprecate the html4 writer see: #938. A printed warning message will come in a different PR. Before this can happen, we have to ensure that the html5 writer is available which was added in sphinx2.

This patch also drops python2 and python versions less than 3.5. Sphinx2 requires these so it is safe to remove support.

Fixes #679