Fix code includes style
Currently this is inheriting a lot of rules from wyrm that dont apply here.
To fix this some wyrm rules are manually overriden or copied.

With all the fixes made in the theme or work around wyrm issues we will want to carefully see what can be removed when we move to bootstrap.

Fixes #930