Fonts: Use fontsource package for loading fonts
fontsource is the new defacto source for google fonts.
Originally, this was `typeface-*` however that is now deprecated.
Not sure why but there does not seem to be woff2 formats for the `all` character sets.
But the new fonts are already smaller than the old woff2 as we now only include the needed font weight.