Update html Layout fixes issues and improves html5 usage
This patch updates the css layout to use new html5 semantic tags. This patch also reworks the sidebar to reduce the needed css/html nesting and makes things simpler. As a result this improves the responsiveness of the sidebar. The sidebar is now always fixed to 300px (this was a bug report will update to add the task).

This patch helps with html5 adoption and makes moving to bootstrap easier. This patch also helps with docutils 0.17 which makes better use of semantic tags.