Fix field list with long terms
Currently, fieldlist extend the full content and may even overflow. Now with this patch, terms are limited to 20% of the field list.

This patch also refactors the margin to be handled by css grid gap property.

Possible future changes:

These are things that I have noticed that can be fixed in future patches.

- Remove left padding -- seems unnecessary to me
- Remove bottom margin on last child globally
- Collapse grid on mobile