Fix failing CI tests
Issue was newer node versions are not compatible with our dev envorment.
This failed because the `cimg/python:3.9-node` defaulted to a newer version of node.
To fix this we can use multiple images to specify a specific node version.

I settled on `node:15.14.0` as it is the last working version.

After we get tests passing we can work on supporting newer node versions.