Fix hidden search bar on all media types
On certain viewport sizes, an increased .wy-header height caused the
‘Github Stars’ and ‘Join Slack’ buttons in .wy-github-stars to overlap
the search bar. These buttons would either partially overlap the search
bar and breadcrumbs, or completely overlap the left hamburger menu,
which contains the search bar when expanded.

This adds a media query to hide the two buttons when the viewport width
is 1169px or less.

Signed-off-by: Stacy Kim <stacy.kim@ucla.edu>